requirements.txt
flask>=2.0.1
python-dotenv>=0.19.0
llama-cpp-python>=0.2.0
diffusers>=0.24.0
torch>=2.0.0
transformers>=4.30.0
accelerate>=0.20.0
safetensors>=0.3.1
requests>=2.28.0
pillow>=9.0.0
numpy>=1.24.0
